+++ .gitignore
@@ -129,7 +129,7 @@ dmypy.json
 .pyre/
 
 # data
-.bin
+*.bin
 
 # pytorch model
 .th
