+++ malt/__init__.py
@@ -1,3 +1,3 @@
 from . import data, models, molecule, trainer, policy, agents, metrics, utils
-from .molecule import Molecule, AssayedMolecule
+from .molecule import Molecule
 from .data.dataset import Dataset
