+++ malt/data/__init__.py
@@ -1,1 +1,1 @@
-from . import dataset, collections, utils
+from . import dataset, collections
