+++ malt/data/tests/test_dataset.py
@@ -1,9 +1,7 @@
 import pytest
 
-
 def test_import():
     from malt.data import dataset
-
 
 def test_build_dataset():
     from malt.data.dataset import Dataset
@@ -52,7 +50,7 @@ def test_dataset_view():
     assert isinstance(g, dgl.DGLGraph)
     assert isinstance(y, torch.Tensor)
     assert y.shape[0] == 2
-    assert y.dim() == 1
+    assert y.dim() == 2
 
 def test_dataset_view_batch_of_g():
     import dgl
@@ -62,33 +60,10 @@ def test_dataset_view_batch_of_g():
     m1 = Molecule("C", metadata={'y': 0.0})
     m2 = Molecule("CC", metadata={'y': 0.0})
     ds = Dataset([m1, m2])
-    _ds = ds.view(batch_size=2, collate_fn="batch_of_g")
+    _ds = ds.view(batch_size=2, by="g")
     g = next(iter(_ds))
     return type(g)
     assert isinstance(g, dgl.DGLGraph)
     assert (
         g.number_of_nodes() == m1.g.number_of_nodes() + m2.g.number_of_nodes()
     )
-
-def test_dataset_with_assayed_molecule_view():
-    import torch
-    import dgl
-    from malt.data.utils import collate_metadata
-    from malt.data.dataset import Dataset
-    from malt.molecule import AssayedMolecule
-
-    m1 = AssayedMolecule("C", metadata={'assay': [{'y': 0.0}]})
-    m2 = AssayedMolecule("CC", metadata={'assay': [{'y': 0.0}]})
-    ds = Dataset([m1, m2])
-    _ds = ds.view(
-        batch_size=2,
-        assay='assay',
-        by=['g', 'y'],
-        batch_meta=collate_metadata
-    )
-    assert isinstance(_ds, torch.utils.data.DataLoader)
-    g, y = next(iter(_ds))
-    assert isinstance(g, dgl.DGLGraph)
-    assert isinstance(y, torch.Tensor)
-    assert y.shape[0] == 2
-    assert y.dim() == 1
